requirements.txt
spidev==3.5
RPi.GPIO==0.7.0
pigpio==1.78